Moved CMS api file
